backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
psycopg2-binary==2.9.9
redis==5.0.4
celery==5.4.0
celery[redis]==5.4.0
smartapi-python==1.4.8
websocket-client==1.8.0
pyotp==2.9.0
logzero==1.7.0
pycryptodome==3.20.0
pandas==2.0.3
numpy==1.26.4
requests==2.31.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.1
pydantic-settings==2.3.1
sendgrid==6.11.0
scikit-learn==1.4.2
xgboost==2.0.3
apscheduler==3.10.4
httpx==0.27.0
feedparser==6.0.11
pytest==8.2.0
pytest-asyncio==0.23.7
